refactor(runtime-pi): collapse to MCP-direct as the only LLM surface
Removes the env-flag gate (`RUNTIME_MCP_CLIENT` /
`RUNTIME_MCP_DIRECT_TOOLS`) and the entire alias layer. The runtime
container now wires `provider_call`, `run_history`, and `llm_complete`
unconditionally via `buildMcpDirectFactories` whenever a sidecar URL
is set; CLI mode (no sidecar) keeps working with no provider tools.

- env.ts: drops `runtimeMcpClient` + `runtimeMcpDirectTools` fields.
- entrypoint.ts: collapses the three-way branch (direct / alias /
  legacy) to a single MCP-direct path. The `buildInContainerBundle`
  helper is hoisted up so phase C can use it as the empty-bundle
  fallback for `buildMcpDirectFactories`.
- extensions/mcp-bridge.ts: deleted (alias layer retired).
- extensions/mcp-result.ts: folded into mcp-direct.ts (single consumer
  remained after the alias layer went away).
- test/mcp-bridge.test.ts + test/provider-tools-wiring.test.ts:
  deleted -- both were testing the deleted code paths exclusively.
